Move restore helper into editor utils
